+++ nacc_attribute_deriver/utils/date.py
@@ -45,7 +45,7 @@ def calculate_age(date1: date | None, date2: date | None) -> Optional[int]:
 
 
 def calculate_interval(date1: date | None, date2: date | None) -> Optional[int]:
-    """Calculates the interval in days between two dates.
+    """Calculates the interval in months between two dates.
 
     Args:
         date1: The earlier date
@@ -56,7 +56,7 @@ def calculate_interval(date1: date | None, date2: date | None) -> Optional[int]:
     if not date1 or not date2:
         return None
 
-    return (date2 - date1).days
+    return (date2.year - date1.year) * 12 + date2.month - date1.month
 
 
 def get_unique_years(dates: List[str]) -> Set[int]: